js: Extract message_edit_history.js.
