+++ .eslintrc.json
@@ -231,6 +231,7 @@
                 "marked": false,
                 "md5": false,
                 "message_edit": false,
+                "message_edit_history": false,
                 "message_events": false,
                 "message_fetch": false,
                 "message_flags": false,
