+++ static/js/global.d.ts
@@ -68,6 +68,7 @@ declare let local_message: any;
 declare let localstorage: any;
 declare let markdown: any;
 declare let message_edit: any;
+declare let message_edit_history: any;
 declare let message_events: any;
 declare let message_fetch: any;
 declare let message_flags: any;
