+++ static/js/message_events.js
@@ -324,7 +324,7 @@ exports.update_messages = function update_messages(events) {
         // Rerender "Message edit history" if it was open to the edited message.
         if ($('#message-edit-history').hasClass('in') &&
             msg.id === parseInt($('#message-history').attr('data-message-id'), 10)) {
-            message_edit.fetch_and_render_message_history(msg);
+            message_edit_history.fetch_and_render_message_history(msg);
         }
     }
